changes for v2.70~9
The tanks GUI changed subsantially in v2.70_9.

The stock code now shrinks the tank tile to 1 line if space is limited. In this mode, the fluid name scrolls.

Blinking has been restored. It indicates the pump is active. This may have always been the case, I just missed that.

Some of the enhancements are now in the stock code.